Fixed ContainerOpenPacket.

+++ src/network/protocol/packet/ContainerOpenPacket.php
@@ -39,7 +39,7 @@ class ContainerOpenPacket extends RakNetDataPacket{
 		$this->reset();
 		$this->putByte($this->windowid);
 		$this->putByte($this->type);
-		$this->putByte($this->slots);
+		$this->putShort($this->slots);
 		$this->putInt($this->x);
 		$this->putInt($this->y);
 		$this->putInt($this->z);
